Revert to old name
